Got code working with mysql and postgresql
