+++ .dockerfile
@@ -1,4 +1,3 @@
-<<<<<<< HEAD:.dockerfile
 # Use an official Python runtime as a base image
 FROM python:3.10-slim
 
@@ -14,31 +13,3 @@ RUN pip install --no-cache-dir -r requirements.txt
 
 # Run app.py when the container launches
 CMD ["python", "app.py"]
-=======
-# Use an official Python runtime as a base image
-FROM python:3.9-slim
-
-# Set the working directory in the container
-WORKDIR /app
-
-# Copy only the requirements file first to leverage Docker cache
-COPY requirements.txt /app/
-
-# Install any needed packages specified in requirements.txt
-RUN pip install --no-cache-dir -r requirements.txt
-
-# Install icalendar
-RUN pip install icalendar
-
-# Copy the current directory contents into the container at /app
-COPY . /app
-
-# Make port 80 available to the world outside this container
-EXPOSE 80
-
-# Define environment variable
-ENV NAME World
-
-# Run app.py when the container launches
-CMD ["python", "app.py"]
->>>>>>> 13ad38db06f5c53efeffad500bbc621db0d79ff5:Dockerfile
