+++ dev-docker-compose.yml
@@ -9,21 +9,45 @@ services:
       - 5000:5000
     volumes:
       - .:/app
+    restart: unless-stopped
     environment:
       FLASK_APP: app.py
       FLASK_ENV: development
       FLASK_DEBUG: 1
-      POSTGRESQL_USERNAME: example
-      POSTGRESQL_PASSWORD: example
-      POSTGRESQL_HOST: http://postgresqldb
-      POSTGRESQL_PORT: 5432
-      POSTGRESQL_DB_NAME: example
+      POSTGRESQL_USERNAME: postgres
+      POSTGRESQL_PASSWORD: postgres
+      POSTGRESQL_HOST: postgresqldb
+      POSTGRESQL_DB_NAME: postgres
+      # MYSQL_USERNAME: user
+      # MYSQL_PASSWORD: password
+      # MYSQL_HOST: mater-mysqldb
+      # MYSQL_DB_NAME: db
+      DB_TYPE: postgresql
   postgresqldb:
     image: postgres
+    container_name: mater-postgres
     restart: always
     ports:
       - 5432:5432
     environment:
-      POSTGRES_PASSWORD: example
-      POSTGRES_USER: example 
-      POSTGRES_DB: example
+      POSTGRES_PASSWORD: postgres
+      POSTGRES_USER: postgres
+      POSTGRES_DB: postgres
+    volumes:
+      - postgresql_volume:/var/lib/postgresql/data
+  # mysqldb:
+  #   container_name: mater-mysqldb
+  #   image: mysql
+  #   restart: always
+  #   environment:
+  #     MYSQL_DATABASE: db
+  #     MYSQL_USER: user
+  #     MYSQL_PASSWORD: password
+  #     MYSQL_ROOT_PASSWORD: password
+  #   ports:
+  #     - 3306:3306
+  #   volumes:
+  #     - mysql_volume:/var/lib/mysql
+volumes:
+  postgresql_volume:
+  # mysql_volume:
