+++ dev.dockerfile
@@ -1,5 +1,8 @@
 # Use an official Python runtime as a base image
-FROM python:3.10-slim
+FROM python:3.10-slim as builder
+
+RUN apt-get update &&  \ 
+    apt-get install -y libpq-dev gcc
 
 # Set the working directory in the container
 WORKDIR /app
@@ -9,7 +12,5 @@ COPY . /app
 
 # Install any needed packages specified in requirements.txt
 RUN pip install --no-cache-dir -r requirements.txt
-
-
 # Run app.py when the container launches
 CMD ["python", "app.py"]
